[Week 3] Workshop Continuation

+++ VGP134_Week3/VGP134_Week3/PetVet.cpp
@@ -7,6 +7,5 @@ void PetVet::DoCheckup(int id)
 	//Pet petVisited;
 	//petVisited = PetRegistry::GetPet(id);
 
-	std::cout << "Checked up " << PetRegistry::Get()->GetPet(id).mName << " (" << PetRegistry::Get()->GetPet(id).mPetType << ")\n";
-	
+	std::cout << "Checked up " << PetRegistry::Get()->GetPet(id).mName << " (" << (int)PetRegistry::Get()->GetPet(id).mPetType << ")\n";
 }
